Move pkg/kubectl/apply and subdirectories to staging
