+++ staging/src/k8s.io/kubectl/pkg/apply/empty_element.go
