+++ staging/src/k8s.io/kubectl/pkg/apply/parse/visitor.go
@@ -17,8 +17,8 @@ limitations under the License.
 package parse
 
 import (
+	"k8s.io/kubectl/pkg/apply"
 	"k8s.io/kubectl/pkg/util/openapi"
-	"k8s.io/kubernetes/pkg/kubectl/apply"
 )
 
 // ItemVisitor provides an interface for Items to Accept and call
