+++ staging/src/k8s.io/kubectl/pkg/apply/strategy/merge_conflict_test.go
@@ -19,7 +19,7 @@ package strategy_test
 import (
 	. "github.com/onsi/ginkgo"
 
-	"k8s.io/kubernetes/pkg/kubectl/apply/strategy"
+	"k8s.io/kubectl/pkg/apply/strategy"
 )
 
 var _ = Describe("Comparing fields of remote and recorded ", func() {
