+++ staging/src/k8s.io/kubectl/pkg/apply/strategy/replace_primitive_list_test.go
@@ -19,7 +19,7 @@ package strategy_test
 import (
 	. "github.com/onsi/ginkgo"
 
-	"k8s.io/kubernetes/pkg/kubectl/apply/strategy"
+	"k8s.io/kubectl/pkg/apply/strategy"
 )
 
 var _ = Describe("Replacing fields of type list with openapi", func() {
