+++ staging/src/k8s.io/kubectl/pkg/apply/strategy/retain_keys_test.go
@@ -19,169 +19,175 @@ package strategy_test
 import (
 	. "github.com/onsi/ginkgo"
 
-	"k8s.io/kubernetes/pkg/kubectl/apply/strategy"
+	"k8s.io/kubectl/pkg/apply/strategy"
 )
 
-var _ = Describe("Merging fields of type list-of-primitive with openapi", func() {
-	Context("where one of the items has been deleted", func() {
-		It("should delete the deleted item", func() {
+var _ = Describe("Merging fields with the retainkeys strategy", func() {
+	Context("where some fields are only defined remotely", func() {
+		It("should drop those fields ", func() {
 			recorded := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "a"
-  - "b"
-  - "c"
+spec:
+  strategy:
 `)
 			local := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "a"
-  - "c"
+spec:
+  strategy:
+    type: Recreate
 `)
 			remote := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "a"
-  - "b"
-  - "c"
+spec:
+  strategy:
+    type: RollingUpdate
+    rollingUpdate:
+      maxUnavailable: 1
+      maxSurge: 1
 `)
 			expected := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "a"
-  - "c"
+spec:
+  strategy:
+    type: Recreate
 `)
 			run(strategy.Create(strategy.Options{}), recorded, local, remote, expected)
 		})
 	})
 
-	Context("where one of the items is only on the remote", func() {
-		It("should move the remote-only item to the end but keep it", func() {
+	Context("where some fields are defined both locally and remotely", func() {
+		It("should merge those fields", func() {
 			recorded := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "a"
-  - "b"
+spec:
+  strategy:
 `)
 			local := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "a"
-  - "b"
+spec:
+  strategy:
+    type: RollingUpdate
+    rollingUpdate:
+      maxUnavailable: 2
 `)
 			remote := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "c"
-  - "b"
-  - "a"
+spec:
+  strategy:
+    type: RollingUpdate
+    rollingUpdate:
+      maxSurge: 1
 `)
 			expected := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "a"
-  - "b"
-  - "c"
+spec:
+  strategy:
+    type: RollingUpdate
+    rollingUpdate:
+      maxUnavailable: 2
+      maxSurge: 1
 `)
 			run(strategy.Create(strategy.Options{}), recorded, local, remote, expected)
 		})
 	})
 
-	Context("where one of the items is repeated", func() {
-		It("should de-duplicate the repeated items", func() {
+	Context("where the elements are in a list and some fields are only defined remotely", func() {
+		It("should drop those fields ", func() {
 			recorded := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "a"
-  - "b"
+spec:
+  template:
+    spec:
 `)
 			local := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "a"
-  - "b"
-  - "a"
+spec:
+  template:
+    spec:
+      volumes:
+      - name: cache-volume
+        emptyDir:
 `)
 			remote := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "a"
-  - "b"
+spec:
+  template:
+    spec:
+      volumes:
+      - name: cache-volume
+        hostPath:
+          path: /tmp/cache-volume
 `)
 			expected := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "a"
-  - "b"
+spec:
+  template:
+    spec:
+      volumes:
+      - name: cache-volume
+        emptyDir:
 `)
 			run(strategy.Create(strategy.Options{}), recorded, local, remote, expected)
 		})
 	})
 
-	Context("where some items are deleted and others are on remote only", func() {
-		It("should retain the correct items in the correct order", func() {
+	Context("where the elements are in a list", func() {
+		It("the fields defined both locally and remotely should be merged", func() {
 			recorded := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "a"
-  - "b"
-  - "c"
+spec:
+  template:
+    spec:
 `)
 			local := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "a"
-  - "c"
-  - "a"
+spec:
+  template:
+    spec:
+      volumes:
+      - name: cache-volume
+        hostPath:
+          path: /tmp/cache-volume
+        emptyDir:
 `)
 			remote := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "d"
-  - "b"
-  - "c"
-  - "a"
-  - "e"
+spec:
+  template:
+    spec:
+      volumes:
+      - name: cache-volume
+        hostPath:
+          path: /tmp/cache-volume
+          type: Directory
 `)
 			expected := create(`
 apiVersion: apps/v1
 kind: Deployment
-metadata:
-  finalizers:
-  - "a"
-  - "c"
-  - "d"
-  - "e"
+spec:
+  template:
+    spec:
+      volumes:
+      - name: cache-volume
+        hostPath:
+          path: /tmp/cache-volume
+          type: Directory
+        emptyDir:
 `)
 			run(strategy.Create(strategy.Options{}), recorded, local, remote, expected)
 		})
