+++ staging/src/k8s.io/kubectl/pkg/apply/strategy/test_swagger.json
